map, chart, and authentication api
