+++ portfolio/src/main/java/com/google/sps/data/Comment.java
@@ -1,5 +1,4 @@
 package com.google.sps.data;
-//edited here
 /* Comment attributes */
 public final class Comment {
   private final long id;
