+++ portfolio/src/main/java/com/google/sps/servlets/LoginServlet.java
@@ -1,0 +1,74 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.sps.servlets;
+import java.util.*; 
+import com.google.gson.Gson;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
+import java.io.IOException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import com.google.appengine.api.datastore.DatastoreService;
+import com.google.appengine.api.datastore.DatastoreServiceFactory;
+import com.google.appengine.api.datastore.Entity;
+import com.google.appengine.api.datastore.PreparedQuery;
+import com.google.appengine.api.datastore.Query;
+import com.google.appengine.api.datastore.Query.SortDirection;
+import com.google.sps.data.User;
+/* handles login/logout mechanism, determining current account state of user */
+@WebServlet("/login")
+public class LoginServlet extends HttpServlet {
+  @Override
+  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    response.setContentType("text/html");
+    UserService userService = UserServiceFactory.getUserService();
+    Dictionary dict = new Hashtable();
+    String json;
+    User user = new User(-1, "null", "-1");
+    if(userService.isUserLoggedIn()) {
+      String userEmail = userService.getCurrentUser().getEmail();
+      String urlToRedirectToAfterUserLogsOut = "/";
+      String logoutUrl = userService.createLogoutURL(urlToRedirectToAfterUserLogsOut);
+      dict.put("login-status", "in");
+      dict.put("userEmail", userEmail);
+      dict.put("url", logoutUrl);
+      Query query = new Query("User");
+      DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
+      PreparedQuery results = datastore.prepare(query);
+      boolean first = true;
+      for(Entity entity : results.asIterable()){
+        String email = (String) entity.getProperty("userEmail");
+        if(userEmail.equals(email)){
+          long id = (long) entity.getKey().getId();
+          String answer = (String) entity.getProperty("answer");
+          user = new User(id, userEmail, answer);
+          first = false;
+        }
+      }
+      dict.put("first-time?", first);
+      dict.put("user", user);
+    }else{
+      String urlToRedirectToAfterUserLogsIn = "/";
+      String loginUrl = userService.createLoginURL(urlToRedirectToAfterUserLogsIn);
+      dict.put("login-status", "out");
+      dict.put("url", loginUrl);
+    }
+    json = new Gson().toJson(dict);
+    response.setContentType("application/json");
+    response.getWriter().println(json);
+  }
+}
