+++ portfolio/src/main/webapp/styles.css
@@ -50,3 +50,15 @@ button:hover{
   height: 600px;
   width: 800px;
 }
+.modal {
+  position: fixed; 
+  z-index: 1; 
+  overflow: auto; 
+  padding-top: 100px; 
+  width: 100%; 
+  height: 100%;
+  left: 0;
+  top: 0;
+  background-color: black; 
+  background-color: rgba(0,0,0,0.3); 
+}
